+++ frontend/src/components/Main/ModelsCheckBoxList.js
@@ -4,60 +4,38 @@ import {inject, observer} from "mobx-react";
 import ModelsCheckBox from "./ModelsCheckBox";
 import ModelsReadOnly from "./ModelsReadOnly";
 import {toJS} from "mobx";
+import ModelsCheckBoxHeader from "./ModelsCheckBoxHeader";
 
-@inject("mainStore")
+@inject("modelsStore")
 @observer
 class ModelsCheckBoxList extends Component {
     render() {
-        const {mainStore} = this.props,
+        const {modelsStore} = this.props,
             handleCheckbox = e => {
                 let model_name = e.target.name;
                 let checked = e.target.checked;
                 let value = e.target.value;
-                mainStore.toggleModelsCheckBox(model_name, checked, value);
+                modelsStore.toggleModelsCheckBox(model_name, checked, value);
             },
-            models = toJS(mainStore.getModelTypeList()),
-            modelsHeaders = mainStore.getmodelsHeaders(),
-            isEditSettings = mainStore.getEditSettings();
+            isEditSettings = modelsStore.getEditSettings();
+        let dataset_type = modelsStore.dataset_type;
+        modelsStore.getModelsList(dataset_type);
+        let models = toJS(modelsStore.models);
+        let modelsHeaders = modelsStore.getmodelsHeaders();
         return (
             <div>
-                {models.length ? (
-                    <table className="modelscheckbox table table-bordered table-sm">
-                        <thead className="table-primary">
-                            {Object.keys(modelsHeaders).map((item, index) => {
-                                return [
-                                    <tr key={index}>
-                                        <th>{modelsHeaders[item].model}</th>
-                                        {modelsHeaders[item].values.map((dev, index) => {
-                                            return [
-                                                <th key={index} colSpan={dev.colspan}>
-                                                    {dev.name}{" "}
-                                                    {(dev.name === "Enable") & isEditSettings ? (
-                                                        <input
-                                                            type="checkbox"
-                                                            name={dev.model_name + "-All"}
-                                                            onChange={handleCheckbox}
-                                                            checked={dev.isChecked}
-                                                        />
-                                                    ) : null}
-                                                    &emsp;
-                                                    {dev.model_name === "bayesian_model_average"
-                                                        ? dev.prior_weight
-                                                        : null}
-                                                </th>,
-                                            ];
-                                        })}
-                                    </tr>,
-                                ];
-                            })}
-                        </thead>
-                        {isEditSettings ? (
-                            <ModelsCheckBox models={models} onChange={handleCheckbox} />
-                        ) : (
-                            <ModelsReadOnly models={models} />
-                        )}
-                    </table>
-                ) : null}
+                <table className="modelscheckbox table table-bordered table-sm">
+                    <ModelsCheckBoxHeader
+                        model_headers={modelsHeaders}
+                        isEditSettings={isEditSettings}
+                        onChange={handleCheckbox}
+                    />
+                    {isEditSettings ? (
+                        <ModelsCheckBox models={models} onChange={handleCheckbox} />
+                    ) : (
+                        <ModelsReadOnly models={models} />
+                    )}
+                </table>
             </div>
         );
     }
