+++ frontend/src/stores/RootStore.js
@@ -2,6 +2,7 @@ import dataStore from "./DataStore";
 import outputStore from "./OutputStore";
 import mainStore from "./MainStore";
 import optionsStore from "./OptionsStore";
+import modelsStore from "./ModelsStore";
 
 class RootStore {
     constructor() {
@@ -9,6 +10,7 @@ class RootStore {
         this.mainStore = mainStore;
         this.outputStore = outputStore;
         this.optionsStore = optionsStore;
+        this.modelsStore = modelsStore;
     }
 }
